Fix Endatix.Framework dependency on Infrastructure and Core
