+++ src/Endatix.Framework/Endatix.Framework.csproj
@@ -14,14 +14,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <InternalsVisibleTo Include="Endatix.Hosting" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <ProjectReference Include="..\Endatix.Core\Endatix.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Ardalis.GuardClauses" />
     <PackageReference Include="Microsoft.FeatureManagement.AspNetCore" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" />
