+++ src/Endatix.Framework/Modules/EndatixModuleRegistration.cs
@@ -7,9 +7,12 @@ namespace Endatix.Framework.Modules;
 /// <summary>
 /// Host-configuration helpers for <see cref="IEndatixModule"/> registration.
 /// </summary>
-internal static class EndatixModuleRegistration
+public static class EndatixModuleRegistration
 {
-    internal static bool ShouldRegister(IConfiguration configuration, IEndatixModule module)
+    /// <summary>
+    /// Returns whether the module should be registered given current host configuration.
+    /// </summary>
+    public static bool ShouldRegister(IConfiguration configuration, IEndatixModule module)
     {
         if (module is not IHasFeatureFlag featureFlagModule)
         {
@@ -19,7 +22,10 @@ internal static class EndatixModuleRegistration
         return IsFeatureFlagEnabled(configuration, featureFlagModule.FeatureFlag);
     }
 
-    internal static bool IsFeatureFlagEnabled(IConfiguration configuration, string featureName)
+    /// <summary>
+    /// Returns whether the named feature flag is enabled in <c>Endatix:FeatureFlags</c>.
+    /// </summary>
+    public static bool IsFeatureFlagEnabled(IConfiguration configuration, string featureName)
     {
         if (string.IsNullOrWhiteSpace(featureName))
         {
